Move all the elsfp files from cmis/ to cmis/elsfp/ and address open review comments
