+++ sonic_platform_base/sonic_xcvr/mem_maps/public/cmis/elsfp/pages/consts.py
@@ -1,0 +1,8 @@
+"""
+    consts.py
+
+    ELSFP-specific CMIS page number constants.
+"""
+
+ELSFP_ADVERTISEMENTS_FLAGS_CTRL_PAGE = 0x1A
+ELSFP_SETPOINTS_MON_PAGE = 0x1B
