+++ sonic_platform_base/sonic_xcvr/mem_maps/public/cmis/pages/consts.py
@@ -19,8 +19,3 @@ TUNABLE_LASER_CTRL_STATUS_PAGE = 0x12
 MODULE_PERF_DIAG_CTRL_PAGE = 0x13
 VDM_ADVERTISING_CTRL_PAGE = 0x2F
 CDB_MESSAGE_PAGE = 0x9F
-
-
-# ELSFP-specific CMIS pages
-ELSFP_ADVERTISEMENTS_FLAGS_CTRL_PAGE = 0x1A
-ELSFP_SETPOINTS_MON_PAGE = 0x1B
